1 test fail, rv64ui-p-st_ld.dump test_50

+++ include/riscv/register_file.h
@@ -14,13 +14,13 @@ public:
 
     void reset();
 
-    void write(u32 index, u32 value);
-    [[nodiscard]] u32 read(u32 index) const;
+    void write(u32 index, u64 value);
+    [[nodiscard]] u64 read(u32 index) const;
 
-    const std::array<u32, REGISTER_COUNT>& raw() const { return registers_; }
+    const std::array<u64, REGISTER_COUNT>& raw() const { return registers_; }
 
 private:
-    std::array<u32, REGISTER_COUNT> registers_{};
+    std::array<u64, REGISTER_COUNT> registers_{};
 };
 
 std::string reg_name(u32 index);
